Implementing (fairly bogus) leader/follower iterators for sparse domains
With the goal of removing the --enable-parallel-iterators flag,
added leader/follower iterators for sparse domains.  Because
it's not entirely clear how zippered parallel iteration should
work on sparse domains (and because we only call forall iterators
on sparse domains in the "solo" context), I made the leader
yield a boolean and the follower do the normal iteration using
a single thread for now.  I also made an overloaded follower
that generates a compilerError if anything other than a bool
is sent in (i.e., if the sparse domain is zippered with anything
else).

The arrayofSpsArray-workaround tests created their own bogus
sparse implementation, so I added leader/follower calls for
them as well.

I think this should take care of our existing sparse tests
as far as the --enable-parallel-iterators flag goes.  Of
course we still need to figure out what leaders/followers
mean in this context (and these specific tests make me want
to see the "solo" iterator case get implemented as well... :)




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14750 3a8e244f-b0f2-452b-bcba-4c88e055c3ca