Refactored the conversion of begin/cobegin/coforall/on blocks into nested functions and the inlining of leader iterators.  Previously we converted begin/cobegin/coforall/on blocks into nested functions right before flattening all nested functions in the lowerIterators pass. Now we convert begin/cobegin/coforall/on blocks into nested function at the beginning of the parallel pass.
Refactored the inlining of leader iterators.  Previously we inlined
leader iterators during function resolution (near the end).  Now we
inline leader iterators at the beginning of the lowerIterators pass
right before where other iterators are inlined, for-loops are expanded
into calls to the iterator methods, and iterators are lowered.

This lets us inline iterators after all other nested functions are
flattened but before begin/cobegin/coforall/on blocks (which can
appear in leader iterators) are turned into functions.

This fixes the last of the tests that did not work with
--enable-parallel-iterators.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14752 3a8e244f-b0f2-452b-bcba-4c88e055c3ca