Suppressing sync warning on the MTA compiler
Suppressing one of two sync warnings on the MTA
compiler, as I've convinced myself that our code
should be fairly well-behaved in this regard.

Unfortunately, there is a second related warning
that the MTA compiler generates that it not
squelchable.  Made a feature request to the
compiler team to enable them to be squashed.

Here's the test code I used to convince myself,
for future reference:

#include <machine/types.h>

typedef struct _S {
  int32_t data;
  int32_t pad;
  int64_t morestuff;
} S;

int main() {
  S myS;

  myS.data = 1;
  myS.pad = 2;
  myS.morestuff = 3;

  purge(&(myS.data));

  writeef(&(myS.data), 4);

  printf("myS is: (%d, %d, %d)\n", myS.data, myS.pad, myS.morestuff);

  int32_t val = readfe(&(myS.data));
  printf("Reading out the data value yields: %d\n", val);

  int32_t five = 5;
  writeef((int64_t*)&(myS.data), five);

  printf("myS is: (%d, %d, %d)\n", myS.data, myS.pad, myS.morestuff);

  val = readfe((int64_t*)&(myS.data));
  printf("Reading out the data value yields: %d\n", val);
}





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14753 3a8e244f-b0f2-452b-bcba-4c88e055c3ca