Adding support for nightly segment=fast testing
Adding a nightly-gasnet.cron script that does our nightly GASNet testing.
(also updated my cron job to run this rather than doing it by-hand as it
was before).  This will test both segment everything and segment fast.

Updated the nightly script to decorate the subject line with the segment
in the case of GASNet.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14760 3a8e244f-b0f2-452b-bcba-4c88e055c3ca