Changed how constructors are searched for
Instead of looking through the list of global types (gTypes), the global list of
functions (gFns) is searched.  A function is considered to be a constructor if
it has the default constructor flag, or if it has the wrapper, inline, and
invisible function flags.  Going through gTypes led to types that were bogus,
and some of these types could not be distinguished from non-bogus ones.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14762 3a8e244f-b0f2-452b-bcba-4c88e055c3ca