Improving user error message for non-cross-mounted filesystems
Improved the expect script used by the PBS launcher to do a
quick test on the executable it's trying to run to see whether
it's accessible from the compute node's file system or not.
If not, a somewhat more helpful error message is printed.
If so, it then executes the program as it used to.

Removed an extraneous log_user 1 command that would seem to
duplicate what already happens with the interact command.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14763 3a8e244f-b0f2-452b-bcba-4c88e055c3ca