Make -v flag help with pbs debugging
Made the log_user=0 command in our expect script only
get generated if the verbose flag is not used on the
executable.  This means that using the -v flag
currently causes all the expect script's commands to
show through, which is nice for debugging.  This sort
of thing may want us to eventually support multiple
levels of verbosity (e.g., a "chapel" level and a
"underlying machinery" level).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14764 3a8e244f-b0f2-452b-bcba-4c88e055c3ca