PBS launcher: made the PBS launcher a bit more qsub-version-sensitive
Different versions of qsub accept different arguments, particularly
in specifying the number of nodes to use.  Started to distinguish
between different versions, though the only two cases we currently
distinguish between is ours and ORNL's (which is hard to distinguish
due to the lack of a supported --version flag; which is exactly what
we look for...).

Haven't quite tested this on our XTs yet, owing to the difficulty
of getting source over there.  About to try after this commit.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14774 3a8e244f-b0f2-452b-bcba-4c88e055c3ca