PBS Launch: un-hard-coded number of cores on an XT
Put in some simple code to dynamically query the number of
cores on an XT, as required to determine the size argument
for the nccs version of qsub (the only place that this
function is currently used).  My implementation is somewhat
naive in that it assumes all nodes have the same number of
cores.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14778 3a8e244f-b0f2-452b-bcba-4c88e055c3ca