Added code to fragment local blocks into smaller local blocks that do not contain control flow including yield and return statements.  This along with the bug fix below makes it so that local blocks can be used in iterators around yield statements.  Removed two futures.
Fixed a bug in handling of local blocks in which wide references to
wide classes would only be narrowed in terms of the wide reference.
This suggests a bug in code generation of narrow references to wide
classes.  I think the right way to fix all of these may be to
dereferences references to wide classes before using them in
GET_MEMBER/SET_MEMBER primitives.  This could potentially simplify
what we have to deal with since we would always dereferences wide
references to wide classes before using them.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14785 3a8e244f-b0f2-452b-bcba-4c88e055c3ca