PBS launcher: added lustre filesystem check
Our existing check to see whether it was legal to launch a
Chapel program on an XT is 100% accurate (it essentially
tries the launch) but potentially slow (you have to sit in
a qsub queue before getting the results.

After some discussions with the XT software team, I added
a test of the cwd's filesystem to see whether it is lustre
or not.  If it is, the old system is executed as it was
before; if not, it warns you that the file system may not
support application launch and/or file I/O as you might
expect and asks whether or not you want to continue.
y, Y, or return continue, n, N, and Ctrl-C abort, and anything
else gives you another chance to answer.  As with other
warnings, this is squelched by the -q flag.

This implementation is conservative in the following senses:

1) there may be file systems from which a launch can occur,
   but which are not truly cross-mounted (thus, I/O won't
   appear transparent).  /tmp is an example.

2) DVS can provide some file reflection apart from lustre.

It should also be noted that most of this logic is very
XT-specific, and not related to qsub/aprun.  As we support
qsub on broader platforms, we will want to specialize
these checks appropriately.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14788 3a8e244f-b0f2-452b-bcba-4c88e055c3ca