Added a new function at the beginning of normalize called markAlignedArrays which looks for arrays that have the same distribution.  Then it marks calls to a toFollower function if the function takes a tuple argument that is made up of arrays that share the same distribution.  The first argument is always marked.  This mark takes the form of a tuple type.  The component types are either Distribution or BaseArray.  Distribution indicates that that the component is aligned.  BaseArray indicates that it is not.  Ideally, this could be done with a tuple of parameters but tuples of parameters are not supported.  Therefore, it is a tuple of types.
If a follower iterator is going over an aligned component, then it
will call the these iterator with a bool parameter called aligned that
is set to true.  The array class (subclass of BaseArray) must define a
param function called supportsAlignedFollower that returns true or the
bool parameter will not be used to call the follower iterator.
BaseArray defines such a function that returns false which is the
default.

Fixed a bug in == and != on types in which a reference would be
considered a distinct type even though references are an internal
mechanism.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14790 3a8e244f-b0f2-452b-bcba-4c88e055c3ca