Adding flush to standalone write() calls
I was wrestling with the console I/O test in this
commit on the XT until I realized that our write()
calls to the console are not flushed by default.
While it seems reasonable not to flush write()s
to general files by default (due to overhead),
flushing the console seems helpful for cases like
this one.  I didn't add a flush to writeln since
printf's of "\n" typically result in a flush.
Let me know if there are any objections to this.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14796 3a8e244f-b0f2-452b-bcba-4c88e055c3ca