Added reference counting to the reindex, slice, rankChange and reallocate functions for both domains and ddata.  reallocate still leaks memory because if the temporary array it creates is destroyed, it can end up freeing memory that is needed later on.  Otherwise, the new test case, LeakedMemory5, would not have been a future.
Changed the default value of the counter in BaseDomain back to zero, and added
code to increment the counter when a new array is created, or when an alias is
created through one of the functions above.

Also added a remove method to the List standard module so when the memory in
an array is freed, this method can be called to have the array removed from
the list of arrays associated with a domain.  (This method is called from
DefaultArithmeticArray's destructor.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14800 3a8e244f-b0f2-452b-bcba-4c88e055c3ca