Added support for forall-expressions.  Previously these expressions were supported as for-expressions.  Use --serial or --serial-forall for those semantics.
Whereas for-expressions are replaced by a single iterator,
forall-expressions are replaced by three iterator functions (the
serial iterator, the leader iterator, and the follower iterator), and
all of these iterator functions are defined in a function that returns
the serial iterator.  The others are called via the toLeader and
toFollower function if they are used in a forall-context.  If they are
used in a for-context, the serial iterator is used.

Added a test case of a forall expression with leader follower semantics.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14807 3a8e244f-b0f2-452b-bcba-4c88e055c3ca