parallelized reductions and promotion
This change is guarded by a flag.  Use --sjd to enable.

Reduction code is now generated by the compiler using a forall loop.
It currently calls the accumulate function inside this loop to
accumulate values into the reduction class instance.  This is
unguarded and not terribly safe.  The next step is to make a local
class that accumulates values and then use a 'combine' function to add
these to the global reduction class instance, guarding the combine
calls by a lock.  Eventually we'll want to do this in log time somehow.

Made promotion result in the creation of three functions to define a
parallel iterator rather than a serial iterator (if the promoted
function returns a value).  This involved the following complications.
First, promotion now creates two classes that are visible to function
resolution, the leader and the follower.  These need to be visible
wherever the promoted function is called so I've made them globally
visible.  Since they need an argument that is a specific iterator
class only created by the promotion, they shouldn't clash anywhere.
Second, the leader and follower versions need to resolve calls to the
leader and follower iterators of what is doing the promotion.  I made
the promoted functions use the instantiationPoint mechanism so that
they can use the visibility of whereever the promotion occurs.  Third,
if the promotion is of a nested function (for example, a forall- or
for-loop), then the _toFollower and _toLeader functions (which invoke
primitives) get arguments that exist in other functions.  To handle
this case, I specialize flatten functions to treat the _toFollower and
_toLeader functions specially.  Basically, they are added to the list
of nested functions as necessary and passed the necessary arguments.

This code is guarded by --sjd because several tests need to be
updated.  They work only when promotions and reductions are serial
because they do not have the necessary leader-follower iterators
defined or they violate some parallel restrictions or they need some
error cases handled.  Here is the breakdown of regressions with this
flag:

  reduction of forall-expression over serial sparse domain iterator dimIter

    npb/cg/bradc/cg-sparse.chpl
    performance/compiler/bradc/cg-sparse-timecomp.chpl

  user-defined FTree iterator is not parallel

    studies/madness/aniruddha/madchap/test_diff.chpl
    studies/madness/aniruddha/madchap/test_likepy.chpl
    studies/madness/aniruddha/madchap/test_showboxes.chpl
    studies/madness/aniruddha/madchap/mytests/par-compress/test_compress.chpl
    studies/madness/aniruddha/madchap/mytests/par-reconstruct/test_reconstruct.chpl
    studies/madness/aniruddha/madchap/mytests/par-refine/test_refine.chpl
    studies/madness/aniruddha/madchap/recordBug/test_likepy.chpl
    studies/madness/common/test_diff.chpl
    studies/madness/common/test_likepy.chpl
    studies/madness/common/test_showboxes.chpl
    studies/madness/dinan/mad_chapel/test_diff.chpl
    studies/madness/dinan/mad_chapel/test_likepy.chpl
    studies/madness/dinan/mad_chapel/test_showboxes.chpl

  reduction over for-expression

    puzzles/deitz/puzzle071018.chpl

  zipper of iterators of different rank (unhandled followers)

    studies/hpcc/HPL/stonea/serial/hplExample2.chpl
    studies/hpcc/HPL/stonea/serial/hplExample3.chpl
    studies/hpcc/HPL/stonea/serial/hplExample4.chpl
    studies/hpcc/HPL/stonea/serial/hplExample5.chpl

  DefaultSparseArray only defines a serial iterator

    release/examples/sparse.chpl

  Block1DDom in BradsBlock1D.chpl only defines a serial iterator

    studies/hpcc/STREAMS/bradc/stream-block1dclass.chpl
    studies/hpcc/STREAMS/bradc/stream-block1dpar-arr.chpl
    studies/hpcc/STREAMS/bradc/stream-block1dpar.chpl

  user-defined iterator I is serial

    studies/sudoku/deitz/sudoku2.chpl
    studies/sudoku/deitz/sudoku_compress2.chpl

  zipper of iterators of different rank (tricky)

    studies/sudoku/deitz/sudoku3.chpl
    studies/sudoku/deitz/sudoku_compress3.chpl
    studies/sudoku/dinan/search/sudoku-reduce4.chpl

  random iterators are serial and/or reduction of for-expression

    trivial/deitz/monte.chpl
    trivial/deitz/monte2.chpl
    trivial/deitz/monte3.chpl
    trivial/diten/pi_monte_carlo.chpl

  zipper parallel iteration does not error-check unbounded ranges

    types/range/diten/badRange2.chpl

  reduction over standard module's list data structure

    types/seq/deitz/test_reduce1.chpl


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14811 3a8e244f-b0f2-452b-bcba-4c88e055c3ca