PBS launch: working around jaguar wallclock time issue
Making all our jaguar chapel jobs currently get supported to
the debug queue in order to temporarily get around the issue
of it requiring a wallclock time limit.  Ultimately, I'm
guessing, we'll want a launcher-specific command-line argument
that allows the user to specify a time limit (and require it
when launching on jaguar...).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14812 3a8e244f-b0f2-452b-bcba-4c88e055c3ca