Added initial support for privatizing arrays and domains.
Note that this is fairly unstable, use --no-privatize to disable.

This checkin implements the privatization optimization almost
exclusively in the modules.  There is a little more compiler work to
do to clean things up, etc., but this is the bulk of the code.  There
is also more work to do in the Block1D distribution to make the
replicated classes more private, as discussed below.

Added a flag --no-privatization to turn this optimization off.  This
works by setting the parameter _privatization to false (true if the
optimization is on), and this parameter is accessed a conditional
statement in the modules.

Almost all of the module changes are in ChapelArray.  I abstracted the
constructors for the _array and _domain records to check if the
array/domain class supports this optimization by defining a parameter
function _supportsPrivatization() to return true; this function
returns false in the base array and domain class.  If it is not
supported, then the constructors for _arrays and _domains are called
directly.  Otherwise, primitives are used to build up a privatized
array of array and domain classes and add one instance to it.  These
new primitives are:

  chpl_numPrivateClasses: returns the number of private classes
    created so far.

  chpl_setPrivateClass: sets the local private class; this is called
    on each locale when a privatized array or domain is created.

  chpl_getPrivateClass: gets a local private class based on an
    integral id.

The construction of private objects is done in a critical section
guarded by a global sync variable called _privatizeLock$, defined in
ChapelArray.

The class on here is set to be the constructed array or domain class.
The other classes are setup using a method called privatize that
returns a new class.

We'll also need a method called reprivatize when changes are made to
arrays or domains, but this is not implemented yet.

The array and domain records now contain two class instances that are
the same in the regular case and different in the privatized case.
This is done so that the type of the class can be kept track of during
compilation.  The compiler should clean this up and remove this extra
field, but that is not done yet.

Implemented privatize and supportsPrivatization methods for Block1D
arrays and domains.  These are not perfect yet in that the global
array class on remote locales still points to the global domain class
on the original locale.

This decreases the number of 'gets' used for a write to a local part
of an array on a remote locale from 10 to 3.  One of the remaining
writes has to do with getting a remote reference to the integral id;
the other two have to do with the global domain pointer from the
privatized array classes not being correctly set.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14845 3a8e244f-b0f2-452b-bcba-4c88e055c3ca