Switching default segment=fast memory from hmm to lea
The general impression is that lea may be the more stable
memory allocator, so I am switching the default to it in
order to get a regression run in this weekend and see how 
it compares.  hmm had a number of regressions due to not 
supporting a realloc call yet.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14846 3a8e244f-b0f2-452b-bcba-4c88e055c3ca