Reverting default non-default memory allocator to heapmm
The problems that we run into using the dlmalloc and segment fast
on our workstations are seemingly very varied and hard to predict.
At least to the extent that I've been able to tell so far... 
Setting GASNET_BACKTRACE to 1, I'm getting a variety of behaviors,
though often it's an abort() being called in the msapce_free()
call.  For now, switching the default memory allocator back to
heapmm in order to get cleaner regression tests until we have
a chance to look at this in a focused way.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14851 3a8e244f-b0f2-452b-bcba-4c88e055c3ca