Optimized coforall ... on ... idiom where an on-statement is directly inside a coforall-loop.  In this case, instead of launching threads that do the remote fork and wait, we do a non-blocking remote fork.
Added FLAG_NON_BLOCKING that can be applied to an on-function that has
FLAG_ON to indicate that it should be non-blocking.

Added PRIMITIVE_BLOCK_ON_NB for use in building non-blocking
on-statements as is now done in buildCoforallLoopStmt.

Removed the polling thread from the GASNet runtime, removed dead code,
and cleaned up the non-blocking fork function.  Also various cleanup.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14855 3a8e244f-b0f2-452b-bcba-4c88e055c3ca