Made 'dom' field in privatized array classes point to the privatized domain classes.  I did this using a primitive in the array class but we can probably clean this up so that the primitive only needs to be used in the ChapelArray module.  In so doing, divided the privatize() method into two, privatize1() and privatize2().
Made it so that global arrays and domains are not put on the heap but
are instead private-broadcasted.

These changes make it so that Block1D global array accesses with
--no-checks do not require any communication if the data being
accessed is local.  If --no-checks is not specified, remote accesses
are required to do bounds checking on the domain over which LocArr is
declared, since this domain lives in the distribution.  This domain
can easily be replicated.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14865 3a8e244f-b0f2-452b-bcba-4c88e055c3ca