Updating distribution-oriented comments
Updated the domain/distribution comments for stream to reflect
the fact that the distribution is now on its separate line.
This actually seems a bit clearer to describe now (esp. given
the redundancy in specifying the index set 1..m twice).
Broke the distributions into their own lines in RA as well
for consistency and updated the comments in a similar fashion.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14867 3a8e244f-b0f2-452b-bcba-4c88e055c3ca