Updates to HPL
Formatted HPL in order to get it a bit closer to the pattern
we've established with our other HPCC benchmarks.  Had to
fix up probSize.chpl because it didn't take into account the
fact that HPL is a 2D problem rather than 1D, so added an
optional "rank" argument to the compute and print ProblemSize
routines to deal with this.  It could use a bit of cleanup,
particularly the case where the problem size is a power of 2
and the matrix is 2d which I didn't bother to handle at all.

Some mysteries:

- I can't seem to make the indexType of hpl.chpl int(64)
  rather than int.  Should do some back of the envelope
  calculations to see whether int(32) is sufficient.  It
  just might be.

- It seems to me that Norm should be part of our standard
  modules -- it's common stuff and it'd be nice to get it
  right once in the libraries.

- The verify results call seems as though it ought to be
  able to take A[MatrixSpace] as an argument rather than
  A and MatrixSpace as two arguments but that doesn't work.
  Oh, I know why -- fillRandom is filling the matrix and
  the vector, not just the matrix.  Hmm... should think
  about this a little more.  I guess the aliases could be
  set up within verifyResults...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14871 3a8e244f-b0f2-452b-bcba-4c88e055c3ca