First stab at HPCC-style fragmented stream implementation
In walking through the HPCC implementation of stream and 
sketching out what a Chapel version of that would look like,
Sam and I remembered that we had a version that was similar
to that as a lead-in to distributions.  Made a copy of that
version to the 2008 HPCC directory and changed the timing
stuff around so that it was more similar to how the HPCC
official version does it.  In particular, they:

* run numTrials versions of the benchmark completely locally
  to each locale, and compute the min, max, and avg times
  across those trials within that locale

* do min, max, and sum reductions over the min, max, and avg
  times (dividinng the sum reduction by the number of locales)
  in order to get some notion of a "global" min, max, and avg,
  where all of those numbers are going to be the style of
  number that a single locale would compute and none of them
  say anything about how the coordinated computation of the
  stream as a whole is done (since the mins, maxs, and avgs
  could have come from different trials by this point, e.g.,
  not to mention the fact that the timings aren't aggregated).

To this end, Sam and I sketched out a version in Chapel that
we believed had the same logic (local reductions within the
coforall; global reductions outside of the coforall) though
we haven't verified that we got it all right yet.  This is
a snapshot of that code -- Sam will verify that it's computing
what we intend for it to.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14875 3a8e244f-b0f2-452b-bcba-4c88e055c3ca