In RA, made the distributions, domains, and arrays into global variables so that privatization works without the worry of passing the integral id in an on-clause.
In RA, introduced an on-clause into the main forall loop.

Improved communication debugging/diagnosis in GASNet so that the 'put'
used for the private broadcast of the debug flag is not reported.

Made .domain efficient on privatized domains.

Introduced a .dist method on domains that returns a domains
distribution.

Made it so that privatization of the Block1D domain class replicates
the distribution class on every locale.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14876 3a8e244f-b0f2-452b-bcba-4c88e055c3ca