This checkin is David's work to enable an optimization in which values that are only read in an on-, begin-, cobegin-, or coforall-block are passed to the function implementing the block by value rather than by reference.  This is particularly important for the on-block because it eliminates remote communication.
I made two additions to the code.  I moved it before heap-allocating data in
the parallel pass.  This didn't require any changes and has the advantage of
eliminating the need to put data on the heap in the inner loop of RA (as well
as eliminating the remote reads of the random number r).  Second, I added
analysis to disable this optimization in the presence of accessing sync or
single variables.

These tests regress:

  multilocale/local/deitz/test_local2

  In this test, we are supposed to halt on a runtime error because of an access
  that is not local, but with this optimization, the access is local.

  multilocale/local/deitz/test_local3

  In this test, it looks like a var function that returns a local is set in an
  on-function, but the variable is not put on the heap because it can't be seen
  there.  This seems independent of this optimization.  The reason it fails now
  is because the read is no longer seen.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14879 3a8e244f-b0f2-452b-bcba-4c88e055c3ca