Updates to stream-fragmented.chpl
Based on conversations with David that stream-fragmented was
not getting competitive performance with stream, rewrote the
timings and GB/s calculations to more closely resemble MPI's
and inserted a local block.  This helps, but on my desktop,
we still seem to be a factor of ~3 off from the GB/s time
reported by stream.chpl.

Possible causes:

1) It looks like some COMM_ calls are not being removed from the main
   loop in spite of the local { } block.

2) Could be some factor related to the fact that we're using the
   DefaultArithmetic distribution rather than Block1D.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14885 3a8e244f-b0f2-452b-bcba-4c88e055c3ca