Simplified communication interface by removing ability to "self-launch"
Since we started supporting GASNet, we've supported an ability in
the communication layer that has permitted an a.out file to self-
launch the numLocales copies of itself, due to the fact that (a)
GASNet had this capability, and (b) we didn't have a launcher
capability yet.  The downsides to this approach have been (i) 
making main.c much more confusing than it needs to be; (ii)
requiring strange calls in the chpl_comm interface; and (iii)
generally mixing launch responsibilities and concerns into the
main binary.  One way in which this has shown up recently has
been in making the setup of memory more complicated than necessary.

For all of these reasons, I implemented a true amudprun launcher
for GASNet and stripped out the "self-launch" capabilities from
the comm interface in general and GASNet specifically.

Changes:

- removed the _chpl_comm_user_invocation() call, which used to
  be used to determine whether main was being called by the user
  or a self-launch loop

- also removed chpl_comm_create_argcv which was only used for
  the self-launch capability

- removed the runInGDB argument from chpl_comm_init because it
  is now called before the arguments have been parsed.  Added
  a chpl_comm_run_in_gdb call to request that the communication
  layer run itself in gdb, if it is capable of doing so.  It
  returns a flag indicating whether or not it can and a status
  if it did.  I believe that in doing this, I've disabled the
  --gdb flag for GASNet, but this can be enabled manually
  currently by setting CHPL_COMM_USE_GDB.  I'll fix this in
  a separate checkin once I'm back at an X-capable machine.

- made amudprun the default launcher for gasnet running over udp
  in the launcher.pl script

- added Makefile variables CHPL_MAKE_TARGET_COMPILER and 
  CHPL_MAKE_HOST_COMPILER to store the target and host
  compilers, because when compiling the launchers (host code), 
  I need to compute the location of the amudprun command based
  on the target.

- Updated the GASNet and ARMCI chpl_cfg_comm.c files to match
  the new interface

- Cleaned up main based on these changes



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14903 3a8e244f-b0f2-452b-bcba-4c88e055c3ca