Removing requirement that CHPL_HOME be set when building launch-amudprun
My previous check-in required CHPL_HOME to be set when building
launch-amudprun which we typically haven't required for the
runtime.  This checkin relaxes that requirement by computing
CHPL_HOME from the current working directory using the CURDIR
variable in gnu Makefiles.  Ultimately, we may want to make the
query of CHPL_HOME be at Chapel compiler-invocation time rather
than the building of the runtime so that the Chapel installation
can be moved around after being built.  I didn't tackle this
here, though.  (Note that going as far as checking CHPL_HOME
at Chapel program execution time is also a possibility that
I prototyped, but this seemed like a poor idea because I don't
think we want to require environment variable to be set to 
execute a Chapel program.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14908 3a8e244f-b0f2-452b-bcba-4c88e055c3ca