Replaced gAsts with component vectors gSymExprs, gCallExprs, gFnSymbols, ... for every (non-abstract) AST subclass
Fixed a few bugs, mostly in scope resolution, that relied on a
specific ordering in gAsts.  This is concerning with regards to
robustness and I believe a useful test would be to randomize the
global vectors and see if compilation still works.

Introduced foreach_ast macro that takes a macro and applies it to
every component vector.

Removed some dead code.

Added a few comments here and there.

Removed extra verification code and flag, as it seems less necessary
now.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14909 3a8e244f-b0f2-452b-bcba-4c88e055c3ca