More fragmented stream versions
Changed our main 2008 stream-fragmented.chpl to be the code
that David did his timings on this week -- using promotion
rather than a zippered forall loop.

Made a second copy of this test without the local block
for performance comparison studies going forward:
stream-fragnolocal.chpl

Copied the previous version with the zippered forall loop
into a new test: stream-fragforall.chpl

Next step will be to add performance files for these, but
I'll probably do that at the same time that I deal with
the issue of copying our 2008 entries into the 
release/examples directory and deal with moving all of
the performance data files around to match.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14919 3a8e244f-b0f2-452b-bcba-4c88e055c3ca