Made timedexec kill child processes more cleanly
Added a setpgrp() so that the child forked by fork will be the
head of its own process group and used a kill -9 in order to
request that the process group be killed, not just the process.
This seems to do a better job of cleaning up runaway gasnet
processes than the previous approach.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14924 3a8e244f-b0f2-452b-bcba-4c88e055c3ca