Added support for conditional function resolution in which a conditional statement around a global "try token" changes the behaviour of function resolution so that the then-part is resolved if possible.  If any error occurs, then it is ignored and the then-part is replaced by the else-part and function resolution continues on it. On the other hand, if no errors occur, then the else-part is removed.
Applied this conditional construct to reductions so that a parallel
reduction is first attempted.  If this fails, then a serial reduction
is attempted.  If this fails, an error message is issued.

Switched the resolvedFns set to a map of FnSymbol* to bool.  This
allows a function that is marked as resolved to be unmarked.  This is
important for conditional function resolution so that a function can
be unmarked if we don't finish evaluating it and then it can be
evaluated later.

Enabled error checking against iteration over an unbounded range in
the range's leader and follower iterators.

This fixes all 35 remaining test cases regressing when the flag
--parallelize-reductions-promotions is used, so I removed it.  These
tests now serialize their reductions.

Next step: Add a warning for when a reduction is serialized.

There is also the case that promotion over a function that does not
return a value is implemented via a forall-loop.  This should be made
to use conditional function resolution so that it can revert to a
for-loop implementation.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14929 3a8e244f-b0f2-452b-bcba-4c88e055c3ca