Moved Norm.chpl to standard modules & refactored standard modules spec section
Moved the Norm.chpl module which we defined in a few places
in the test directory into the standard modules.  It seems
like good functionality, and while our current implementation
could be improved, that's true of many of our standard modules.
The motivating factor in this case is that we'd like to use
it for our HPCC HPL benchmark's verification step.

Rewrote tests that used their own private Norm module to use
the standard Norm module except for test/spec/marybeth which
is an older version that I couldn't get working with the
standard Norm module easily (and didn't try any harder than
that).

Updated the spec to describe the Norm module and refactored
the standard modules into two sections as we discussed before
-- those that are automatically included in a Chapel program
and those that have to be explicitly included.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14935 3a8e244f-b0f2-452b-bcba-4c88e055c3ca