Make chpl_realloc call chpl_malloc if it is passed NULL in heapmm.
Also, made it an error to call allocation/deallocation functions if the heap
is not initialized yet. Previously, it fell back to using the regular malloc
functions. This is no longer required because the launchers use the
mem-default implementation and the _real binaries should never allocate
anything until the heap is initialized.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14944 3a8e244f-b0f2-452b-bcba-4c88e055c3ca