Nested coforall in leader iterator bug
This is an interesting bug (at least I believe it's a bug) that
showed up while developing a "multiple tasks per locale" implementation
of the block iterator.  When using a nested coforall within a leader
iterator, we seem to get a race in which some of the iterations of the
inner coforall spill state into the others, causing redundant executions
and/or executions to be skipped.  Collapsing these coforalls into a
single loop seems to make the problem go away.  It could be that 
something (like the loop index?) is not getting privatized properly?

nestedForall2 was an evolution of the code that probably shouldn't work
(unless we permit parallelism within iterators that are inlineable)
but in any case causes a core dump today.

nestedForall3 demonstrates that if the nested coforall is pulled out
of the leader iterator, it seems to work.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14948 3a8e244f-b0f2-452b-bcba-4c88e055c3ca