Fixed recent gasnet regressions when freeing heap-allocated variables
Before, heap-allocated variables were not freed only if they were passed in to
a function created as a result of a begin, cobegin, or coforall statement.  The
problem was that if an on statement immediately contained a begin statement,
a different kind of function was created.  Now, a heap-allocated variable is
not freed if it is passed to an on function that immediately has a begin
statement inside of it, i.e., if the function has the non blocking flag.

Freeing of heap-allocated variables is enabled by default (with no way to
disable it).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14959 3a8e244f-b0f2-452b-bcba-4c88e055c3ca